requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Deep Learning

keras>=3.0.0

# Data processing
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Logging and utilities
tqdm>=4.65.0

# Optional: For advanced visualizations
plotly>=5.14.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0